fetching runner photo from url
